Refactor structure of partners page
