+++ src/pages/partners/components/index.js
@@ -1,0 +1,2 @@
+export { default as NpoPartnersContent } from './NpoPartnersContent';
+export { default as PartnersContent } from './PartnersContent';
